requirements.txt
# ===== Web Framework =====
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ===== Database =====
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# ===== Validation =====
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# ===== Authentication =====
python-jose[cryptography]==3.3.0

# ===== HTTP Client =====
httpx==0.28.1

# ===== Scheduler =====
apscheduler==3.10.4

# ===== Image Processing =====
Pillow==11.1.0

# ===== Testing =====
pytest==8.3.4
pytest-asyncio==0.24.0

# ===== Utilities =====
python-dotenv==1.0.1
